finish the app

+++ src/users/users.service.ts
@@ -22,7 +22,32 @@ export class UsersService {
       });
     }
 
-    return this.prisma.user.create({ data: createUserDto });
+    const user = await this.prisma.user.create({ data: createUserDto });
+
+    const preceptors = await this.prisma.user.findMany({
+      where: { role: 'PRECEPTOR' },
+    });
+
+    const roleTranslation = {
+      ADMIN: 'Administrador',
+      PRECEPTOR: 'Preceptor',
+    };
+
+    const notificationLink =
+      createUserDto.role === 'ADMIN' ? '/admins' : '/preceptors';
+
+    for (const preceptor of preceptors) {
+      await this.prisma.notification.create({
+        data: {
+          title: `Nuevo ${roleTranslation[createUserDto.role]}`,
+          body: `Se ha agregado a ${createUserDto.firstName} ${createUserDto.lastName}`,
+          userId: preceptor.id,
+          link: `/school/${notificationLink}`,
+        },
+      });
+    }
+
+    return user;
   }
 
   async createUserOnStudent(createUserOnStudent: CreateUserOnStudentDto) {
@@ -35,6 +60,25 @@ export class UsersService {
 
     if (!user) {
       user = await this.prisma.user.create({ data: createUserDto });
+
+      const student = await this.prisma.student.findUnique({
+        where: { id: studentId },
+        select: { classroomId: true },
+      });
+
+      const classroom = await this.prisma.classroom.findUnique({
+        where: { id: student.classroomId },
+        select: { userId: true },
+      });
+
+      await this.prisma.notification.create({
+        data: {
+          title: 'Nuevo Tutor',
+          body: `Se ha agregado a ${createUserDto.firstName} ${createUserDto.lastName}`,
+          userId: classroom.userId,
+          link: `/school/classrooms/${student.classroomId}/students/${studentId}/tutors`,
+        },
+      });
     }
 
     return this.prisma.userOnStudent.create({
@@ -79,6 +123,17 @@ export class UsersService {
     });
   }
 
+  findStudentByUserId(userId: string) {
+    return this.prisma.userOnStudent.findMany({
+      where: { userId },
+      select: {
+        student: {
+          include: { attendance: true },
+        },
+      },
+    });
+  }
+
   async update(id: string, updateUserDto: UpdateUserDto) {
     const { email } = updateUserDto;
     const existingEmail = await this.prisma.user.findFirst({
@@ -89,6 +144,18 @@ export class UsersService {
       throw new ConflictException({
         status: 'exists',
         message: 'Ya existe un usuario con el mismo email',
+      });
+    }
+
+    const accounts = await this.prisma.account.findMany({
+      where: { userId: id },
+    });
+
+    if (accounts.length > 0) {
+      throw new ConflictException({
+        status: 'update',
+        message:
+          'No se puede modificar el email de un usuario con cuentas asociadas',
       });
     }
 
@@ -121,7 +188,28 @@ export class UsersService {
     return deleteResult;
   }
 
-  removeMultiple(ids: string[]) {
+  async removeMultiple(ids: string[]) {
+    // Verificar la cantidad de administradores
+    const adminCount = await this.prisma.user.count({
+      where: { role: 'ADMIN' },
+    });
+
+    // Contar cuántos administradores están en la lista de ids a eliminar
+    const adminsToDelete = await this.prisma.user.count({
+      where: {
+        id: { in: ids },
+        role: 'ADMIN',
+      },
+    });
+
+    // Si se intenta eliminar a todos los administradores, lanzar una excepción
+    if (adminCount === adminsToDelete) {
+      throw new ConflictException({
+        status: 'error',
+        message: 'No se puede eliminar a todos los administradores',
+      });
+    }
+
     return this.prisma.user.deleteMany({ where: { id: { in: ids } } });
   }
 }
